Create GoogleTasksCredential model w/ tests
