+++ app/models/google_tasks_credential.rb
@@ -1,0 +1,3 @@
+class GoogleTasksCredential < Credential
+  include GoogleApp
+end
