+++ test/fixtures/credentials.yml
@@ -24,6 +24,16 @@ keith_gmail:
   properties: {"expires_at":1717004605,"expires":true,"scope":"https://www.googleapis.com/auth/gmail.modify https://www.googleapis.com/auth/userinfo.email"}
   last_authenticated_at: 2023-12-19 08:40:05
 
+keith_google_tasks:
+  user: keith
+  type: GoogleTasksCredential
+  external_id: gmail_id123
+  oauth_email: keith@hostedgpt.com
+  oauth_token: ya01.a0AXooCgvPlObOgrUgug8OYnvKRssrp0rkzssxxcpUj4Dnj_XEsQ9CBRKaawiqVIyNbl6D67NnkXdYT7uWtrj1qmWG-s0nqeT-zbCE5oT8GHO3XObhjihHzVDhJUKDGTPNqnnj0xErmeyeECzmpdn8WmoSwkn8k2LLoE-6aCgYKAfASARMSFQHGX2MiIXAIUTQ9uJrqXU9Chsj5Cg0171
+  oauth_refresh_token: 1//01l-TCXjNaaTpCgYIARAAGA8SNwF-L9Irko7Aae4EYUBASxpshYXnGWXlGfByGfFmuSpNSbTghckfAewMp4ZZYjcx8y18Chbtjzk
+  properties: {"expires_at":1717004605,"expires":true,"scope":"https://www.googleapis.com/auth/gmail.modify https://www.googleapis.com/auth/userinfo.email"}
+  last_authenticated_at: 2023-12-19 08:40:05
+
 rob_password:
   user: rob
   type: PasswordCredential
